mcp-explorer-requirements.txt
anthropic
httpx
mcp[cli]
fastapi
uvicorn
pyinstaller
pyyaml
rich
art
prompt_toolkit
pycryptodome